+++ fireflight/src/App.js
@@ -5,8 +5,7 @@ import { Route, Redirect } from "react-router-dom";
 import Home from "./components/Home";
 import Dashboard from "./components/Dashboard";
 import AuthForms from "./components/AuthForms/AuthForms";
-// import Address from "./components/Address";
-import Address2 from "./components/Address2";
+import Address from "./components/Address";
 import LandingPage from "./components/LandingPage";
 
 import { GlobalContext } from "./context/contextProvider";
@@ -163,9 +162,7 @@ function App() {
             />
           )}
         />
-
-        {/* <Route path="/address" component={Address} /> */}
-        <Route path="/address" component={Address2} />
+        <Route path="/address" component={Address} />
       </AddressContext>
     </AppWrapper>
   );
