+++ fireflight/src/styles/App.scss
@@ -17,7 +17,7 @@
 @import "model";
 @import "model-splits";
 @import "dashboard";
-@import "address2";
+@import "address-form";
 
 html {
 	font-size: 62.5%;
